// ClientApp/src/app/services/dts-api.ts
export interface TaskItemResponseDto {
    /** The description of the task. */
    description?: string | undefined;
    /** The due date of the task. */
    dueDate?: Date;
    /** The Id of the Task. */
    id?: number;
    /** The status of the task. */
    status?: TaskItemStatus;
    /** The title of the task. */
    title?: string;
}

/** Enum representing the status of a TaskItem. */
export enum TaskItemStatus {
    Completed = 0,
    Created = 1,
    InProgress = 2,
}

/** Represents the data transfer object for creating a TaskItem. */
export interface CreateTaskItemCommandDto {
    /** The description of the task. */
    description?: string | undefined;
    /** The due date of the task. */
    dueDate?: Date;
    /** The status of the Task Item. */
    status?: TaskItemStatus;
    /** The title of the task. */
    title?: string;
}